Final Commit
Carry Ripple Addition or subtraction requires similar circuitry, the
only difference is at the input of the elements, so let us consider the
structure of a Carry ripple adder and then we can relate i[t to the
subtraction.
A carry ripple adder makes an addition of two binary numbers and gives
out two outputs which are sum and carry out, so when these carry ripple
blocks are aligned in series so as to perform a multi bit function the
carry out from one block goes as a carry in to the next block. And thus
the carry out in the last bit is the actual carry out.
The carry ripple subtraction is done in the similar way but, say for
example we are trying to perform the function between two numbers a and
b so instead of feeding the number b directly we give in the 2’s
compliment of it and the regular addition is performed and thus
subtraction is achieved. We try to design this circuit in VHDL